Remove default value for startDate and endDate
The previous default value led to some problems that if no new data was
available, suddenly no data would be shown. The new default is to simply
returned the latest available data if no start of end date is given.

This means that the WHERE clause of the SQL is now dynamically build.